#52 ticket optional feature complete

+++ src/emc-metalnx-shared/src/main/java/com/emc/metalnx/controller/TicketController.java
@@ -45,6 +45,8 @@ import com.emc.metalnx.core.domain.entity.DataGridTicket;
 import com.emc.metalnx.core.domain.exceptions.DataGridConnectionRefusedException;
 import com.emc.metalnx.core.domain.exceptions.DataGridTicketException;
 import com.emc.metalnx.core.domain.exceptions.DataGridTicketNotFoundException;
+import com.emc.metalnx.core.domain.exceptions.UnsupportedDataGridFeatureException;
+import com.emc.metalnx.services.interfaces.ConfigService;
 import com.emc.metalnx.services.interfaces.TicketService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -61,8 +63,15 @@ public class TicketController {
 	@Autowired
 	private LoggedUserUtils loggedUserUtils;
 
+	@Autowired
+	private ConfigService configService;
+
 	@RequestMapping(value = "/", method = RequestMethod.GET)
-	public String index() throws DataGridConnectionRefusedException {
+	public String index() throws DataGridConnectionRefusedException, UnsupportedDataGridFeatureException {
+		if (!configService.getGlobalConfig().isTicketsEnabled()) {
+			logger.error("tickets are not enabled");
+			throw new UnsupportedDataGridFeatureException("tickets disabled");
+		}
 		logger.info("Get tickets page");
 		return "tickets/tickets";
 	}
@@ -70,7 +79,13 @@ public class TicketController {
 	@RequestMapping(value = "/ticketForm", method = RequestMethod.GET)
 	public String createTicketForm(final Model model,
 			@RequestParam(value = "ticketstring", required = false) final String ticketString)
-			throws DataGridConnectionRefusedException, DataGridTicketNotFoundException {
+			throws DataGridConnectionRefusedException, DataGridTicketNotFoundException,
+			UnsupportedDataGridFeatureException {
+
+		if (!configService.getGlobalConfig().isTicketsEnabled()) {
+			logger.error("tickets are not enabled");
+			throw new UnsupportedDataGridFeatureException("tickets disabled");
+		}
 
 		DataGridTicket ticket;
 
@@ -87,15 +102,21 @@ public class TicketController {
 
 	/**
 	 * Finds all tickets in the grid.
-	 * 
+	 *
 	 * @return List of tickets in JSON
 	 * @throws DataGridConnectionRefusedException
 	 *             if Metalnx cannot connect to the grid
+	 * @throws UnsupportedDataGridFeatureException
 	 */
 	@RequestMapping(value = "/", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
 	@ResponseBody
-	public String findAll() throws DataGridConnectionRefusedException, JsonProcessingException {
+	public String findAll()
+			throws DataGridConnectionRefusedException, JsonProcessingException, UnsupportedDataGridFeatureException {
 		logger.info("Find all tickets");
+		if (!configService.getGlobalConfig().isTicketsEnabled()) {
+			logger.error("tickets are not enabled");
+			throw new UnsupportedDataGridFeatureException("tickets disabled");
+		}
 		List<DataGridTicket> tickets = ticketService.findAll();
 
 		Map<String, Object> ticketsAsJSON = new HashMap<>();
@@ -106,26 +127,36 @@ public class TicketController {
 
 	/**
 	 * Finds a specific ticket in the grid by its id or string
-	 * 
+	 *
 	 * @param ticketId
 	 *            ticket id or string
 	 * @return Ticket as JSON
 	 * @throws DataGridConnectionRefusedException
 	 *             if Metalnx cannot connect to the grid
+	 * @throws UnsupportedDataGridFeatureException
 	 */
 	@RequestMapping(value = "/{ticketid}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
 	@ResponseBody
 	public ResponseEntity<DataGridTicket> find(@PathVariable("ticketid") final String ticketId)
-			throws DataGridConnectionRefusedException, DataGridTicketNotFoundException {
+			throws DataGridConnectionRefusedException, DataGridTicketNotFoundException,
+			UnsupportedDataGridFeatureException {
 		logger.info("Find ticket by its ID or String");
+		if (!configService.getGlobalConfig().isTicketsEnabled()) {
+			logger.error("tickets are not enabled");
+			throw new UnsupportedDataGridFeatureException("tickets disabled");
+		}
 		DataGridTicket dgTicket = ticketService.find(ticketId);
 		return new ResponseEntity<>(dgTicket, HttpStatus.OK);
 	}
 
 	@RequestMapping(value = "/{ticketId}", method = RequestMethod.DELETE)
 	public ResponseEntity<String> deleteTicket(@PathVariable final String ticketId)
-			throws DataGridConnectionRefusedException {
+			throws DataGridConnectionRefusedException, UnsupportedDataGridFeatureException {
 		logger.info("Delete ticket by its ID or String");
+		if (!configService.getGlobalConfig().isTicketsEnabled()) {
+			logger.error("tickets are not enabled");
+			throw new UnsupportedDataGridFeatureException("tickets disabled");
+		}
 		boolean ticketDeleted = ticketService.delete(ticketId);
 
 		if (!ticketDeleted) {
@@ -138,8 +169,12 @@ public class TicketController {
 	@RequestMapping(value = "/", method = RequestMethod.DELETE, consumes = MediaType.APPLICATION_JSON_VALUE)
 	@ResponseStatus(HttpStatus.NO_CONTENT)
 	public ResponseEntity<String> bulkDeleteTickets(@RequestBody final List<String> ticketStrings)
-			throws DataGridConnectionRefusedException {
+			throws DataGridConnectionRefusedException, UnsupportedDataGridFeatureException {
 		logger.info("Delete tickets of user: {}", loggedUserUtils.getLoggedDataGridUser().getUsername());
+		if (!configService.getGlobalConfig().isTicketsEnabled()) {
+			logger.error("tickets are not enabled");
+			throw new UnsupportedDataGridFeatureException("tickets disabled");
+		}
 		boolean ticketsDeleted = ticketService.bulkDelete(ticketStrings);
 		if (!ticketsDeleted) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -151,8 +186,12 @@ public class TicketController {
 	@ResponseStatus(value = HttpStatus.CREATED)
 	@ResponseBody
 	public DataGridTicket createTicket(@RequestBody final DataGridTicket ticket)
-			throws DataGridConnectionRefusedException, DataGridTicketException {
+			throws DataGridConnectionRefusedException, DataGridTicketException, UnsupportedDataGridFeatureException {
 		logger.info("Create new ticket");
+		if (!configService.getGlobalConfig().isTicketsEnabled()) {
+			logger.error("tickets are not enabled");
+			throw new UnsupportedDataGridFeatureException("tickets disabled");
+		}
 		ticket.setOwner(loggedUserUtils.getLoggedDataGridUser().getUsername());
 		ticketService.create(ticket);
 		return ticket;
@@ -161,17 +200,49 @@ public class TicketController {
 	@RequestMapping(value = "/", method = RequestMethod.PUT)
 	@ResponseStatus(value = HttpStatus.NO_CONTENT)
 	public void modifyTicket(@RequestBody final DataGridTicket ticket)
-			throws DataGridConnectionRefusedException, DataGridTicketException {
+			throws DataGridConnectionRefusedException, DataGridTicketException, UnsupportedDataGridFeatureException {
 		logger.info("Modify ticket");
+		if (!configService.getGlobalConfig().isTicketsEnabled()) {
+			logger.error("tickets are not enabled");
+			throw new UnsupportedDataGridFeatureException("tickets disabled");
+		}
 		ticketService.modify(ticket);
 	}
 
 	@RequestMapping(value = "/validatehost", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
 	@ResponseBody
 	public HostInfo validateTicketHostname(@RequestParam("hostname") final String hostname)
-			throws UnknownHostException {
+			throws UnknownHostException, UnsupportedDataGridFeatureException {
 		logger.info("Validating ticket hostname {}", hostname);
+		if (!configService.getGlobalConfig().isTicketsEnabled()) {
+			logger.error("tickets are not enabled");
+			throw new UnsupportedDataGridFeatureException("tickets disabled");
+		}
 		return new HostInfo(hostname);
+	}
+
+	public TicketService getTicketService() {
+		return ticketService;
+	}
+
+	public void setTicketService(final TicketService ticketService) {
+		this.ticketService = ticketService;
+	}
+
+	public LoggedUserUtils getLoggedUserUtils() {
+		return loggedUserUtils;
+	}
+
+	public void setLoggedUserUtils(final LoggedUserUtils loggedUserUtils) {
+		this.loggedUserUtils = loggedUserUtils;
+	}
+
+	public ConfigService getConfigService() {
+		return configService;
+	}
+
+	public void setConfigService(final ConfigService configService) {
+		this.configService = configService;
 	}
 }
